불필요코드 삭제 및 컨벤션 통일
