--- src/main/java/com/deare/backend/api/setting/dto/response/UpgradeMembershipResponseDTO.java
@@ -1,11 +1,0 @@
-package com.deare.backend.api.setting.dto.response;
-
-import com.deare.backend.domain.setting.entity.enums.MembershipPlan;
-
-import java.time.LocalDateTime;
-
-public record UpgradeMembershipResponseDTO(
-        MembershipPlan membershipPlan,
-        boolean isPlus,
-        LocalDateTime updatedAt
-) {}
